+++ src/views/helpers/getData.js
@@ -1,162 +1,17 @@
-// const search = require('./postData.js');
-// eslint is expecting 'import' instead of 'require'
-// const app = require('../../app.js');
-// const express = require('express');
-// const cors = require('cors');
-// const morgan = require('morgan');
-const fetch = require('node-fetch');
-// const path = require('path');
-// const bodyParser = require('body-parser');
-
-const Twitter = require('twitter');
-// const handlebars = require('express-handlebars');
-
-require('dotenv').config();
-
-// const app = express();
-
-// body parse will parse the body
-// app.use(bodyParser.urlencoded({ extended: true }));
-
-// app.use(morgan('tiny'));
-// app.use(cors());
-
-//***serving css and views folders contained in public***
-// app.use(
-//     express.static(path.join(__dirname, '..', '/public'), { maxAge: '30d' })
-// );
-// ***** serving homepage index.html
-// app.get('/', (req, res) => {
-//     res.sendFile(path.join(__dirname, '..', '/public/', 'index.html'));
-// });
-
-// ** YOUTUBE **
-//hardcoded youtube username
-const youtubeKey = process.env.GOOGLE_API_KEY;
-// youtube base url
-const youtubeBaseUrl = 'https://www.googleapis.com/youtube/v3/';
-// youtube api 1
-const userNameUrl = 'search?part=snippet&type=channel&maxResults=1&q=';
-// youtube api 2
-const channelUrl = 'channels?part=contentDetails&id=';
-// youtube api 3
-const videoListUrl = 'playlistItems?part=snippet&playlistId=';
-
-// let search = 'metallica';
-
-//const search = postData(); 
-
-
-let tweetsArr = [];
-let arr = [];
-
-// ** TWITTER **
-
-const client = new Twitter({
-    consumer_key: process.env.TWITTER_CONSUMER_KEY,
-    consumer_secret: process.env.TWITTER_CONSUMER_SECRET,
-    access_token_key: process.env.TWITTER_ACCESS_TOKEN_KEY,
-    access_token_secret: process.env.TWITTER_ACCESS_TOKEN_SECRET
-});
-
-const handleTweets = search => {
-    // username = 'heyMrBoi';
-    const params = { screen_name: search, count: 2 };
-    // const tweetsArr = [];
-
-    client.get('statuses/user_timeline', params, (error, tweets) => {
-        if (error) {
-            console.log('tweet error', error);
-        } else {
-            tweets.forEach(item => tweetsArr.push(item.text));
-            // console.log(tweetsArr);
-            arr.push(tweetsArr);
-            return tweetsArr;
-        }
-        // console.log(response.json()); // Raw response object.
-        // res.send(response);
-    });
-};
-
-// end twitter
-
-
+const youtubeApi = require('./youtubeApi');
+const handleTweets = require('./handleTweets');
 
 const getData = (search,cb) => {
-    // if(error) {
+    // const arr = [];
+    // arr.push(youtubeApi(search, cb))
+    const ytresult = youtubeApi(search, cb)
+    // ytresult('anything', ).then(result => console.log(result))
+    console.log("111111111 youtubeApi:", ytresult);
+    // arr.push(handleTweets(search))
+    // console.log("handleTweets:",handleTweets(search));
 
-    // }
-    // app.get('/videos', () => {
-    // console.log(req.body.search);
-    console.log('search in GET', search);
-    return cb(null, () => {
-        fetch(`${youtubeBaseUrl}${userNameUrl}${search}&key=${youtubeKey}`)
-            .then(response => response.json())
-            .then(data => {
-                const channelId = data.items[0].id.channelId;
-                console.log('channel id', channelId);
-                return fetch(
-                    `${youtubeBaseUrl}${channelUrl}${channelId}&key=${youtubeKey}`
-                );
-            })
-            .then(response => response.json())
-            .then(data => {
-                const playListId =
-                    data.items[0].contentDetails.relatedPlaylists.uploads;
-                console.log('this is playListId ', playListId);
-                return fetch(
-                    `${youtubeBaseUrl}${videoListUrl}${playListId}&key=${youtubeKey}`
-                );
-            })
-            .then(response => response.json())
-            .then(data => {
-                const videoId = []; //array of videoId
-                for (let i = 0; i < 4; i++) {
-                    videoId.push(data.items[i].snippet.resourceId.videoId);
-                }
-                console.log(videoId);
-                return videoId;
-            })
-            .catch(error => {
-                console.log('youtube error ', error);
-            })
-            // twitter
-            .then(handleTweets(search))
-            .then(videoId => {
-                console.log('inside final then');
-                // let arr = [];
-                arr.push(videoId);
-
-                // arr.concat(videoId);
-                // arr.push(tweetsArr);
-                console.log(arr);
-                return arr;
-            })
-            .catch(error => {
-                console.log('twitter error ', error);
-            });
-    });
+    // console.log("getData arr:",arr);
+    // return arr;
 };
-
-// const notFound = (req, res, next) => {
-//     res.status(404);
-//     const error = new Error('Not Found!!!');
-//     next(error);
-// };
-
-// const errorHandler = (error, res) => {
-//     res.status(res.statusCode || 500);
-//     res.json({
-//         message: error.message
-//     });
-// };
-
-// app.use(notFound);
-// app.use(errorHandler);
-
-// const port = process.env.PORT || 5000;
-// app.listen(port, () => {
-//     console.log('Listening on port', port);
-// });
 
 module.exports = getData;
